+++ CHANGELOG.md
@@ -9,6 +9,11 @@ This project adheres to [Semantic Versioning](https://semver.org).
 [Commits](https://github.com/thebigmunch/tbm-utils/compare/2.2.1...master)
 
 
+### Added
+
+* Support singular and plural item labels for ``LabelList``.
+
+
 ## [2.2.1](https://github.com/thebigmunch/tbm-utils/releases/tag/2.2.1) (2020-02-25)
 
 [Commits](https://github.com/thebigmunch/tbm-utils/compare/2.2.0...2.2.1)